+++ libs/common/src/lib/guards/set-agent.guard.ts
@@ -1,10 +1,22 @@
 import { inject } from '@angular/core';
-import { ActivatedRouteSnapshot } from '@angular/router';
-import { FeaturesService } from '@flaps/core';
+import { ActivatedRouteSnapshot, Router } from '@angular/router';
+import { FeaturesService, SDKService } from '@flaps/core';
+import { of } from 'rxjs';
 
 export const setAgentGuard = (route: ActivatedRouteSnapshot) => {
+  const sdk: SDKService = inject(SDKService);
+  const router: Router = inject(Router);
   const features = inject(FeaturesService);
   const isRetrievalAgentsEnabled = features.unstable.retrievalAgents;
+
+  const zone = route.paramMap.get('zone');
+  const agentSlug = route.paramMap.get('agent');
+
+  if (!agentSlug || !zone) {
+    return of(router.createUrlTree(['/select']));
+  }
+
+  // this.nuclia.options.zone = zone;
   //TODO agent guard with real permissions
-  return isRetrievalAgentsEnabled;
+  return !sdk.nuclia.options.standalone && isRetrievalAgentsEnabled;
 };